Windows desktop port and ESP-IDF on Windows
First-class Windows support for both the desktop binary (Winsock socket layer, MSVC /W4 /WX clean, .zip release artifact, CI build-windows job) and the ESP32 firmware build (UTF-8 locale shim, IDF venv layout, toolchain PATH including flat Windows tool dirs). A fresh Windows machine now flows git clone -> winget install -> MoonDeck -> Setup ESP-IDF -> Build end-to-end.

tick:873/714/251/247/127/77/2/129/4/229us(FPS:1145/1400/3984/4048/7874/12987/500000/7751/250000/4366) | ESP32:1067KB

ESP32 tick is absent from the one-liner because no ESP32 was wired to the Windows build host during this commit; gate 6 was verified via a clean esp32 firmware build, and the KPI ESP32 section reports flash/DRAM but not a live tick. A future hardware-attached run on Windows (or a fresh monitor.log from any other host) will refill it.

### Core
- platform_desktop.cpp: Winsock port via #ifdef _WIN32 plus thin shims at the top (sock/close_sock/open_sock/make_nonblocking/sockWouldBlock; the would-block check accepts both WSAEWOULDBLOCK and WSAETIMEDOUT). writeChunks toggles FIONBIO around WSASend to emulate MSG_DONTWAIT. A static RAII g_winsockInit calls WSAStartup(2,2) at library load so the test binaries get Winsock too, and fsWriteAtomic uses _fileno/_commit instead of fileno/fsync on Windows.
- main_desktop.cpp: signal() with one-shot semantics on Windows (no sigaction, no SIGPIPE/SIGBUS). _write(2, ...) replaces unistd ::write(STDERR_FILENO, ...) in the crash-handler safeWrite path. Added <cstdint> for uint16_t (was pulled in transitively via unistd.h on POSIX).
- Control.cpp / SystemModule.h: explicit static_casts for the narrowing sites MSVC catches at /W4 that GCC -Wall misses (legitimate, not noise).
- JsonSink.h: portable MM_PRINTF_FORMAT(fmt, va) macro that expands to __attribute__((format(printf))) on GCC/Clang and to nothing on MSVC, which previously parsed the bare attribute as an unknown override specifier and cascaded into 20+ false errors.

### UI
- embed_ui.cmake: now takes a PYTHON_CMD parameter so each caller picks the interpreter (desktop passes \${UV_EXECUTABLE};run;python, ESP32 passes \${Python3_EXECUTABLE}). Windows MSVC build hosts don't ship python3 on PATH, and ESP-IDF docker doesn't have uv -- the parameter avoids both traps.

### Scripts / MoonDeck
- build_desktop.py: --config Release on Windows so MSVC multi-config picks the right build type (CMAKE_BUILD_TYPE is ignored there).
- build_esp32.py: Windows venv layout (Scripts/python.exe), idf_env toolchain PATH walks both bin/-subdir layouts and flat ones (ninja/ccache/idf-exe live at the version-dir root on Windows), PYTHONUTF8=1 + PYTHONIOENCODING=utf-8 injected into idf env, idf_cmd routes through _idf_win_shim.py on Windows.
- _idf_win_shim.py (new): sets the C locale to en_US.UTF-8 via locale.setlocale BEFORE running idf.py via runpy.run_path. Necessary because IDF's _valid_unicode_config uses locale.getlocale() which Windows Python initialises to the system codepage (cp1252 on Dutch Windows etc.) regardless of PYTHONUTF8. sys.path is prepended with the idf tools dir so idf.py's sibling python_version_checker import resolves under runpy.
- setup_esp_idf.py: detects host, runs install.bat on Windows (shell=True) and install.sh elsewhere. Also runs git submodule update --init --recursive --depth 1 idempotently -- Windows users typically do a depth-1 clone which skips submodules, and install.bat then fails.
- package_desktop.py: utf-8 encoding for the README write (cp1252 rejects -> and -- on Windows).
- scenario/run_scenario.py: looks for mm_scenarios.exe in both test/Release/ and test/ (MSVC multi-config vs Ninja single-config).
- check_specs.py, collect_kpi.py, moondeck.py: encoding=\"utf-8\" on read_text() sites that bit on cp1252. collect_kpi.py also passes encoding=\"utf-8\" to subprocess capture (idf.py size emits box-drawing chars that crash cp1252).

### Tests
- unit_FilesystemModule_persistence.cpp: explicit f.close() before remove_all in the two tests that hold an std::ifstream open across the rmtree -- POSIX unlinks open files, Windows blocks.
- unit_MoonModule_movechild.cpp: <string> include -- MSVC STL doesn't pull it in via doctest.h, so std::string == const char* fails to find the doctest comparison overload.
- scenario_runner.cpp: .string().c_str() on std::filesystem::path because Windows path::c_str() returns wchar_t*.
- scenarios/*.json (10 files): auto-written observed[pc-windows] blocks from the scenario runner's drift-tracking persistence.

### Docs / CI
- CLAUDE.md: new \"Use \`uv\` for every Python invocation\" process rule with the ESP-IDF docker exception called out (IDF uses find_package(Python3) and its own venv, not uv).
- docs/building.md: Windows prereqs block (winget install, Developer PowerShell, MSVC multi-config binary location) and an ESP-IDF Windows section (PowerShell git clone, core.longpaths true, IDF version corrected from \"v6.0.0\" which didn't exist on upstream to \"v6.1-dev\" which matches CI).
- README.md: per-OS bullet list for the desktop downloads (macOS tar.gz + Gatekeeper note, Windows zip + SmartScreen note).
- CMakeLists.txt: find_program(UV_EXECUTABLE), ws2_32 link on Windows, _CRT_SECURE_NO_WARNINGS family for cl.exe's deprecated-standard-library warnings, /wd4702 /wd4244 /wd4267 (MSVC-only warnings GCC -Wall doesn't catch; kept the strict /WX otherwise).
- esp32/main/CMakeLists.txt: find_package(Python3 REQUIRED) and \${Python3_EXECUTABLE} for build_info.h and embed_ui -- works on Windows IDF venv (python.exe) and POSIX (python3) without per-host #ifdef.
- .github/workflows/release.yml: new build-windows job on windows-latest, both desktop jobs install uv via astral-sh/setup-uv@v3 and invoke package_desktop.py via uv run, release job's needs + Pages-stage download + final files block all include projectMM-*.zip alongside projectMM-*.tar.gz.
- .gitignore: /dist/ added so CI is the only authoritative producer of release artifacts (a stale dist/projectMM-windows-x64-vRC2.zip was unstaged in the process).

KPI Details:
  Desktop:
    Lights: 16,384
    Binary: 432 KB
    [doctest] test cases:  239 |  239 passed | 0 failed | 0 skipped
    tick: 873us, 714us, 251us, 247us, 127us, 77us, 2us, 129us, 4us, 229us (FPS: 1145, 1400, 3984, 4048, 7874, 12987, 500000, 7751, 250000, 4366) (per scenario)
    === 11 scenario(s), 11 passed, 0 failed ===
    Platform boundary: PASS
    Specs: Spec check: 28 modules, 28 ok, 0 missing, 0 outdated
  ESP32:
    Image: 1,199,889 bytes (35% partition free)
    Flash (code+data): 1067 KB
    DRAM: 38,312 / 180,736 (142,424 free)
  Code:
    64 source files (11642 lines)
    42 test files (6333 lines)
    41 specs, 11 scenarios
    Lizard: 0 warnings

+++ src/core/Control.cpp
@@ -194,7 +194,7 @@ ApplyResult applyControlValue(const ControlDescriptor& c,
             // Password parses identically to Text — only serialization differs.
             // c.max is the buffer size; parseString writes up to maxLen-1 then
             // NUL-terminates, so passing c.max gives "fill the buffer".
-            uint8_t maxLen = c.max > 0 ? c.max : 16;
+            uint8_t maxLen = static_cast<uint8_t>(c.max > 0 ? c.max : 16);
             mm::json::parseString(json, key, static_cast<char*>(c.ptr), maxLen);
             return ApplyResult::Ok;
         }
